Code refactor. Add fix for Android 10. Update README
