+++ README.md
@@ -2,8 +2,9 @@
 Sample illustrating the use of modern Android architectures and components.
 
 Interesting articles and repositories:
-* https://proandroiddev.com/android-viewmodel-dependency-injection-with-kodein-249f80f083c9
-* https://github.com/Karumi/kodein-sample-testing
-* https://https://github.com/pabloogc/FakeShop/
+* Basic Kodein ViewModel injection: https://proandroiddev.com/android-viewmodel-dependency-injection-with-kodein-249f80f083c9
+* Kodein testing sample: https://github.com/Karumi/kodein-sample-testing
+* Basic Flux implementation: https://https://github.com/pabloogc/FakeShop/
+* Mock Kotlin classes in tests: https://proandroiddev.com/mocking-androidtest-in-kotlin-51f0a603d500
