+++ app/src/main/kotlin/com/bq/arch_example/domain/moves/MoveActions.kt
@@ -1,7 +1,7 @@
 package com.bq.arch_example.domain.moves
 
-import com.bq.arch_example.network.MoveName
-import com.bq.arch_example.network.PokemonMove
+import com.bq.arch_example.network.common.MoveName
+import com.bq.arch_example.network.common.PokemonMove
 import mini.BaseAction
 import mini.Task
